refactor(vmm): move block to block/file
In order to allow creation of other block device subtypes,
we move the existing host-file-backed block device
implementation from block to block/file directory.
