+++ src/vmm/src/devices/virtio/block/file/io/async_io.rs
@@ -10,8 +10,8 @@ use logger::log_dev_preview_warning;
 use utils::eventfd::EventFd;
 use utils::vm_memory::{mark_dirty_mem, GuestAddress, GuestMemory, GuestMemoryMmap};
 
-use crate::devices::virtio::block::io::UserDataError;
-use crate::devices::virtio::block::IO_URING_NUM_ENTRIES;
+use crate::devices::virtio::block::file::io::UserDataError;
+use crate::devices::virtio::block::file::IO_URING_NUM_ENTRIES;
 use crate::io_uring::operation::{Cqe, OpCode, Operation};
 use crate::io_uring::restriction::Restriction;
 use crate::io_uring::{IoUring, IoUringError};
