+++ src/vmm/src/devices/virtio/block/file/request.rs
@@ -7,17 +7,18 @@
 
 use std::convert::From;
 
-use logger::{error, IncMetric, METRICS};
+use log::error;
+use logger::{IncMetric, METRICS};
 use utils::vm_memory::{ByteValued, Bytes, GuestAddress, GuestMemoryError, GuestMemoryMmap};
 pub use virtio_gen::virtio_blk::{
     VIRTIO_BLK_ID_BYTES, VIRTIO_BLK_S_IOERR, VIRTIO_BLK_S_OK, VIRTIO_BLK_S_UNSUPP,
     VIRTIO_BLK_T_FLUSH, VIRTIO_BLK_T_GET_ID, VIRTIO_BLK_T_IN, VIRTIO_BLK_T_OUT,
 };
 
-use super::super::DescriptorChain;
+use super::super::super::DescriptorChain;
 use super::{io as block_io, BlockError, SECTOR_SHIFT};
-use crate::devices::virtio::block::device::DiskProperties;
-use crate::devices::virtio::SECTOR_SIZE;
+use crate::devices::virtio::block::file::device::DiskProperties;
+use crate::devices::virtio::file::SECTOR_SIZE;
 use crate::rate_limiter::{RateLimiter, TokenType};
 
 #[derive(Debug, derive_more::From)]
@@ -656,7 +657,7 @@ mod tests {
     use proptest::prelude::*;
     use proptest::strategy::{Map, Strategy, TupleUnion};
 
-    use crate::devices::virtio::block::test_utils::RequestDescriptorChain;
+    use crate::devices::virtio::block::file::test_utils::RequestDescriptorChain;
 
     // Implements a "strategy" for producing arbitrary values of RequestType.
     // This can also be generated by a derive macro from `proptest_derive`, but the crate
