+++ src/vmm/src/version_map.rs
@@ -10,7 +10,7 @@ use semver::Version;
 use versionize::{VersionMap, Versionize};
 
 use crate::device_manager::persist::DeviceStates;
-use crate::devices::virtio::block::persist::BlockState;
+use crate::devices::virtio::block::file::persist::BlockState;
 use crate::devices::virtio::net::persist::NetConfigSpaceState;
 use crate::devices::virtio::QueueState;
 use crate::persist::VmInfo;
